Move scripts under tests/system/test-scripts
