+++ web/package.json
@@ -132,14 +132,14 @@
     "push": "cross-env NODE_ENV=production npm run prod:build && sdk-upload",
     "styleguide": "styleguidist --config webpack/styleguide.config.js server",
     "save-credentials": "sdk-save-credentials",
-    "smoke-test": "bash ./tests/system/smoke-test.sh",
+    "smoke-test": "bash ./tests/system/test-scripts/smoke-test.sh",
     "test": "cross-env NODE_ENV=test jest",
     "test:all": "npm run lint && npm test -- --coverage",
-    "test:build-size": "node tests/system/verify-build-size.js",
+    "test:build-size": "node tests/system/test-scripts/verify-build-size.js",
     "test:e2e": "cross-env NODE_ENV=test node ./node_modules/nightwatch/bin/runner.js -c ./tests/system/nightwatch-config.js --suiteRetries 1",
-    "test:pwa-prod": "bash ./scripts/lighthouse-prod.sh",
-    "test:pwa-ci": "bash ./scripts/lighthouse-ci.sh",
-    "test:pwa-local": "bash ./scripts/lighthouse-local.sh",
+    "test:pwa-prod": "bash ./tests/system/test-scripts/lighthouse/lighthouse-prod.sh",
+    "test:pwa-ci": "bash ./tests/system/test-scripts/lighthouse/lighthouse-ci.sh",
+    "test:pwa-local": "bash ./tests/system/test-scripts/lighthouse/lighthouse-local.sh",
     "test:server": "bash ./scripts/test-server.sh",
     "test:watch": "npm test -- --watch",
     "prod:build": "bash ./scripts/build.sh",
