+++ web/tests/system/test-scripts/lighthouse/lighthouse-prod.sh
@@ -1,0 +1,14 @@
+#!/usr/bin/env bash
+
+OUTPUT_PATH=tests/system/test-scripts/lighthouse/audit-prod
+# See package.json's siteUrl key.
+URL=${1-$npm_package_siteUrl}
+
+lighthouse \
+	--view \
+	--output json \
+	--output html \
+	--output-path $OUTPUT_PATH \
+	$URL
+
+node tests/system/test-scripts/lighthouse/check-score.js
